requirements-ml.txt
torch>=1.9,<3.0
open-clip-torch>=2.20.0
faiss-cpu>=1.7.4
Pillow>=9.0
numpy>=1.21
tqdm>=4.60
requests